Implement generic driver
