+++ Sources/AWSDriver/AWSDriver.swift
@@ -1,8 +1,7 @@
-import Foundation
-
 import AWSSignatureV4
-import Vapor
 import HTTP
+import TLS
+import Vapor
 
 struct CredentialResponse: Codable {
     enum Success: String, Codable {
@@ -23,47 +22,27 @@ struct CredentialResponse: Codable {
     }
 }
 
-public enum Error: Swift.Error {
-    case roleError // Problem getting IAM role credentials
-    case noRolesPresent // This host does not have an associated Instance profile
-}
-
-struct InstanceProfile {
-    func queryLocalEndpoint(urlPath: String) throws -> String {
-        let client = try EngineClientFactory.init().makeClient(hostname:
-            "169.254.169.254", port: 80, securityLayer: .none, proxy: nil)
-        let version = HTTP.Version(major: 1, minor: 1)
-        let request = HTTP.Request(method: Method.get, uri: urlPath, version: version, headers: [:], body: Body.data(Bytes([])))
-        let response = try client.respond(to: request)
-        guard let bytes = response.body.bytes else {
-            throw Error.roleError
-        }
-        return bytes.makeString()
-    }
-
-    func findInstanceRole() throws -> String {
-        let role = try queryLocalEndpoint(urlPath: "http://169.254.169.254/latest/meta-data/iam/security-credentials/")
-        if role.count < 1 {
-            throw Error.noRolesPresent
-        }
-        return role
-    }
-
-    func generateIAMCreds() throws -> (String, String, String?) {
-        let role = try findInstanceRole()
-        let credString = try queryLocalEndpoint(urlPath: "http://169.254.169.254/latest/meta-data/iam/security-credentials/\(role)")
-        let decoder = JSONDecoder()
-        let credentials = try! decoder.decode(CredentialResponse.self, from: credString.data(using: .utf8)!)
-        return (credentials.AccessKeyId, credentials.SecretAccessKey, credentials.Token)
-    }
-}
-
 public struct AWSDriver {
+    public enum Error: Swift.Error {
+        case invalidResponse(Status)
+    }
+
     public let accessKey: String
     public let secretKey: String
     public let token: String?
+    let service: String
+    let host: String
+    let signer: AWSSignatureV4
 
-    public init(accessKey: String? = nil, secretKey: String? = nil, token: String? = nil) throws {
+    public init(service: String, region: Region? = nil, accessKey: String? = nil, secretKey: String? = nil, token: String? = nil) throws {
+        self.service = service
+        var serviceRegion =  Region.usEast1
+        if let awsRegion = region {
+            self.host = "\(self.service).\(awsRegion.rawValue).amazonaws.com"
+            serviceRegion = awsRegion
+        } else {
+            self.host = "\(self.service).amazonaws.com"
+        }
         if let access = accessKey, let secret = secretKey {
             self.accessKey = access
             self.secretKey = secret
@@ -71,5 +50,50 @@ public struct AWSDriver {
         } else {
             (self.accessKey, self.secretKey, self.token) = try InstanceProfile().generateIAMCreds()
         }
+        self.signer = AWSSignatureV4(
+            service: self.service,
+            host: self.host,
+            region: serviceRegion,
+            accessKey: self.accessKey,
+            secretKey: self.secretKey,
+            token: self.token
+        )
+    }
+}
+
+extension AWSDriver {
+    func submitRequest(baseURL: String, query: String, method: HTTP.Method) throws -> String {
+
+        let headers = try signer.sign(path: "/", query: query)
+        let client = try EngineClientFactory().makeClient(hostname: host, port: 443, securityLayer: .tls(Context.init(.client)), proxy: nil)
+
+        print("\(baseURL)/?\(query)")
+        print(headers)
+        let version = HTTP.Version(major: 1, minor: 1)
+        let request = HTTP.Request(method: Method.get, uri: "\(baseURL)/?\(query)", version: version, headers: headers, body: Body.data(Bytes([])))
+        let response = try client.respond(to: request)
+
+        guard response.status == .ok else {
+            print("Response error: \(response)")
+            guard let bytes = response.body.bytes else {
+                throw Error.invalidResponse(response.status)
+            }
+
+            throw try ErrorParser.parse(bytes)
+        }
+
+        guard let bytes = response.body.bytes else {
+            throw Error.invalidResponse(.internalServerError)
+        }
+
+        return bytes.makeString()
+    }
+
+    public func get(baseURL: String, query: String) throws -> String {
+        return try submitRequest(baseURL: baseURL, query: query, method: .get)
+    }
+
+    public func post(baseURL: String, query: String) throws -> String {
+        return try submitRequest(baseURL: baseURL, query: query, method: .post)
     }
 }
